Introduce CrayPackage and apply it to cray_* packages
Replaced the generic Package base class with CrayPackage
in Cray-specific packages to consolidate redundant logic,
enforce consistent metadata, and simplify maintenance.
