+++ repos/spack_repo/builtin/packages/cray_fftw/package.py
@@ -1,13 +1,12 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-from spack_repo.builtin.build_systems.generic import Package
+from spack_repo.builtin.build_systems.cray import CrayPackage
 
 from spack.package import *
 
 
-class CrayFftw(Package):
+class CrayFftw(CrayPackage):
     """FFTW is a C subroutine library for computing the discrete Fourier
     transform (DFT) in one or more dimensions, of arbitrary input
     size, and of both real and complex data (as well as of even/odd
@@ -20,7 +19,6 @@ class CrayFftw(Package):
     """
 
     homepage = "https://support.hpe.com/"
-    has_code = False  # Skip attempts to fetch source that is not available
 
     maintainers("haampie", "lukebroskop")
 
@@ -41,14 +39,6 @@ class CrayFftw(Package):
     variant("openmp", default=False, description="Enable OpenMP support.")
     variant("mpi", default=True, description="Activate MPI support")
     depends_on("mpi", when="+mpi")
-
-    def install(self, spec, prefix):
-        raise InstallError(
-            self.spec.format(
-                "{name} is not installable, you need to specify "
-                "it as an external package in packages.yaml"
-            )
-        )
 
     @property
     def libs(self):
