+++ repos/spack_repo/builtin/packages/cray_libsci/package.py
@@ -1,17 +1,16 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-from spack_repo.builtin.build_systems.generic import Package
+from spack_repo.builtin.build_systems.cray import CrayPackage
 
 from spack.package import *
 
 
-class CrayLibsci(Package):
+class CrayLibsci(CrayPackage):
     """The Cray Scientific Libraries package, LibSci, is a collection of
     numerical routines optimized for best performance on Cray systems."""
 
     homepage = "https://docs.nersc.gov/development/libraries/libsci/"
-    has_code = False  # Skip attempts to fetch source that is not available
 
     version("23.02.1.1")
     version("22.11.1.2")
@@ -26,9 +25,6 @@ class CrayLibsci(Package):
     version("16.07.1")
     version("16.06.1")
     version("16.03.1")
-
-    conflicts("platform=windows")
-    conflicts("platform=darwin")
 
     variant("shared", default=True, description="enable shared libs")
     variant("openmp", default=False, description="link with openmp")
@@ -50,7 +46,7 @@ class CrayLibsci(Package):
 
     @property
     def modname(self):
-        return "cray-libsci/{0}".format(self.version)
+        return f"cray-libsci/{self.version}"
 
     @property
     def external_prefix(self):
@@ -95,11 +91,3 @@ class CrayLibsci(Package):
     @property
     def libs(self):
         return self.blas_libs
-
-    def install(self, spec, prefix):
-        raise InstallError(
-            self.spec.format(
-                "{name} is not installable, you need to specify "
-                "it as an external package in packages.yaml"
-            )
-        )
