+++ repos/spack_repo/builtin/packages/cray_mpich/package.py
@@ -4,20 +4,19 @@
 
 import os
 
+from spack_repo.builtin.build_systems.cray import CrayPackage
 from spack_repo.builtin.build_systems.cuda import CudaPackage
-from spack_repo.builtin.build_systems.generic import Package
 from spack_repo.builtin.build_systems.rocm import ROCmPackage
 from spack_repo.builtin.packages.mpich.package import MpichEnvironmentModifications
 
 from spack.package import *
 
 
-class CrayMpich(MpichEnvironmentModifications, Package, CudaPackage, ROCmPackage):
+class CrayMpich(MpichEnvironmentModifications, CrayPackage, CudaPackage, ROCmPackage):
     """Cray's MPICH is a high performance and widely portable implementation of
     the Message Passing Interface (MPI) standard."""
 
     homepage = "https://docs.nersc.gov/development/compilers/wrappers/"
-    has_code = False  # Skip attempts to fetch source that is not available
 
     maintainers("etiennemlb", "haampie")
 
@@ -41,10 +40,6 @@ class CrayMpich(MpichEnvironmentModifications, Package, CudaPackage, ROCmPackage
     depends_on("cray-pmi")
     depends_on("libfabric")
 
-    depends_on("c", type="build")
-
-    requires("platform=linux", msg="Cray MPICH is only available on Cray")
-
     # cray-mpich 8.1.7: features MPI compiler wrappers
     variant("wrappers", default=True, when="@8.1.7:", description="enable MPI wrappers")
 
@@ -60,7 +55,7 @@ class CrayMpich(MpichEnvironmentModifications, Package, CudaPackage, ROCmPackage
 
     @property
     def modname(self):
-        return "cray-mpich/{0}".format(self.version)
+        return f"cray-mpich/{self.version}"
 
     @property
     def external_prefix(self):
@@ -101,14 +96,6 @@ class CrayMpich(MpichEnvironmentModifications, Package, CudaPackage, ROCmPackage
             spec.mpicxx = spack_cxx
             spec.mpifc = spack_fc
             spec.mpif77 = spack_f77
-
-    def install(self, spec, prefix):
-        raise InstallError(
-            self.spec.format(
-                "{name} is not installable, you need to specify "
-                "it as an external package in packages.yaml"
-            )
-        )
 
     @property
     def headers(self):
